poll: parse json data into Model structs
as the values may be null we cannot parse using this form:

    let foo: Foo = serde_json::from_value(bar)

manual_connect is the exception to this as right now we are just parsing
dummy data which is never null.

otherwise if the data is empty we simply write empty values to NodeInfo.
this will be rendered as Null in ui::ui()