dnetview: major refactor of Model/ underlying data structure
we create 3 types of info objects, NodeInfo, ConnectInfo, and
SessionInfo. NodeInfo is the parent and ConnectInfo and SessionInfo are children.

These info objects are passed into an enum:

enum SelectableObject {
    Node(NodeId)
    Session(SessionId)
    Connect(ConnectId)
}

We protect this data using mutexes and write it to a hashmap:

Mutex<FxHashMap<u32, SelectableObject>

View has not yet been updated with this new design.