outbound_session: create run2() method that changes run() behavior to new whitelist protocol.
get an address from the whitelist using whitelist_fetch_address_with_lock().

try to establish a connection to it. on success, update the last_seen field for the address on the whitelist.

TODO: we're still using try_connect here which quarantines a peer on failure to connect. In the greylist update we will simply drop this connection and move on.