store: don't remove from greylist or whitelist on anchorlist upgrade
because of the refinery running in the background, if we remove a peer
from the white or greylist on upgrade it can create an index error in
when seperate threads execute this code at the same time:

refinery:
    upgrades node to whitelist, removes from greylist

upgrade_host:
    upgrades node to anchorlist, removes from greylist

leaving a "safe" peer on the grey or whitelist is not a problem. the
only impact is that we risk selecting a peer from the whitelist that we
are already connected as an anchor, but p2p checks exist to protect
from this.

equally, if we remove from the greylist or whitelist on upgrade_host this can
happen:

upgrade_host:
    upgrades node to anchorlist, removes from greylist

protocol_addr:
    recv Addr
    do we have this node? no, add to greylist
    refinery: promotes to whitelist, etc

the above scenario makes removing the host in this case redundant.