net: make a generic base class for PeerDiscovery
Making PeerDiscovery generic means that in the future we can have
different types of PeerDiscovery.

Specifically, swarming requires us to have 2 kinds of Peer Discovery, a
an overlay network PeerDiscovery (like the one we have currency), and a
PeerDiscovery for subnets that looks for subnets within the overlay
network.

This paves the way for future changes of this nature