protocol_version: bugfix
The Error result of the version exchange was not being dealt with.
Instead we were only handling when the version exchange timed out.

There was also a bug in the handling of the failure in session/mod.rs
where the error case was similarly ignored. We now explicitly return
with an error in the case that the version exchange fails or times out.