net: fix race condition provoked by channel.ban()
If a node we are connected to as part of a refinery session sends
arbitrary messages, it could trigger a "ban", which would result in a
panic if the refinery was moving the peer to whitelist at the same time
we are sending the node to the blacklist.

Following this commit, we ignore messages without dispatchers if they
come from a refine session.